Remove `adminMeta` argument to `getAdminMeta` function in custom field types and the associated types for it
