+++ .changeset/poor-dogs-exercise.md
@@ -1,0 +1,5 @@
+---
+'@keystone-6/core': major
+---
+
+Removes `adminMeta` argument to `getAdminMeta` function in custom field types and the associated types for it
