+++ packages/core/src/admin-ui/system/generateAdminUI.ts
@@ -4,9 +4,10 @@ import fs from 'fs-extra';
 import resolve from 'resolve';
 import { GraphQLSchema } from 'graphql';
 import { Entry, walk as _walk } from '@nodelib/fs.walk';
-import type { KeystoneConfig, AdminMetaRootVal, AdminFileToWrite } from '../../types';
+import type { KeystoneConfig, AdminFileToWrite } from '../../types';
 import { writeAdminFiles } from '../templates';
 import { serializePathForImport } from '../utils/serializePathForImport';
+import { AdminMetaRootVal } from './createAdminMeta';
 
 const walk = promisify(_walk);
